composefs/gc: GC bootloader binaries
Instead of deleting bootloader binaries in `delete_deployment` function,
add logic to GC them along with everything else. So our source or truth
becomes the bootloader entries which is also what we use for getting the
status.

The flow now becomes when deleting a deployment:
- Delete the bootloader entries (.conf files and user.cfg for grub)
- GC the bootloader binaries + images + state dir + composefs objects

This way the GC remains idempotent and we handle in-between deletion
failures gracefully.

Also, this gives us the opportunity to retrieve the deleted bootloader
entries using the data (binaries) from the /boot or /esp in the future


refactor: Add helper methods

Add `require_esp` method to Storage struct
Add `list_all_deployments` method to Host struct

+++ crates/lib/src/bootc_composefs/delete.rs
@@ -2,11 +2,10 @@ use std::{io::Write, path::Path};
 
 use anyhow::{Context, Result};
 use cap_std_ext::{cap_std::fs::Dir, dirext::CapStdExtDirExt};
-use composefs_boot::bootloader::{EFI_ADDON_DIR_EXT, EFI_EXT};
 
 use crate::{
     bootc_composefs::{
-        boot::{BootType, SYSTEMD_UKI_DIR, find_vmlinuz_initrd_duplicates, get_efi_uuid_source},
+        boot::{BootType, get_efi_uuid_source},
         gc::composefs_gc,
         rollback::{composefs_rollback, rename_exchange_user_cfg},
         status::{get_composefs_status, get_sorted_grub_uki_boot_entries},
@@ -22,7 +21,11 @@ use crate::{
 };
 
 #[fn_error_context::context("Deleting Type1 Entry {}", depl.deployment.verity)]
-fn delete_type1_entry(depl: &DeploymentEntry, boot_dir: &Dir, deleting_staged: bool) -> Result<()> {
+fn delete_type1_conf_file(
+    depl: &DeploymentEntry,
+    boot_dir: &Dir,
+    deleting_staged: bool,
+) -> Result<()> {
     let entries_dir_path = if deleting_staged {
         TYPE1_ENT_PATH_STAGED
     } else {
@@ -32,15 +35,6 @@ fn delete_type1_entry(depl: &DeploymentEntry, boot_dir: &Dir, deleting_staged: b
     let entries_dir = boot_dir
         .open_dir(entries_dir_path)
         .context("Opening entries dir")?;
-
-    // We reuse kernel + initrd if they're the same for two deployments
-    // We don't want to delete the (being deleted) deployment's kernel + initrd
-    // if it's in use by any other deployment
-    let should_del_kernel = match depl.deployment.boot_digest.as_ref() {
-        Some(digest) => find_vmlinuz_initrd_duplicates(digest)?
-            .is_some_and(|vec| vec.iter().any(|digest| *digest != depl.deployment.verity)),
-        None => false,
-    };
 
     for entry in entries_dir.entries_utf8()? {
         let entry = entry?;
@@ -68,7 +62,6 @@ fn delete_type1_entry(depl: &DeploymentEntry, boot_dir: &Dir, deleting_staged: b
                 // Boot dir in case of EFI will be the ESP
                 tracing::debug!("Deleting EFI .conf file: {}", file_name);
                 entry.remove_file().context("Removing .conf file")?;
-                delete_uki(&depl.deployment.verity, boot_dir)?;
 
                 break;
             }
@@ -85,12 +78,6 @@ fn delete_type1_entry(depl: &DeploymentEntry, boot_dir: &Dir, deleting_staged: b
                 tracing::debug!("Deleting non-EFI .conf file: {}", file_name);
                 entry.remove_file().context("Removing .conf file")?;
 
-                if should_del_kernel {
-                    delete_kernel_initrd(&bls_config.cfg_type, boot_dir)?;
-                } else {
-                    tracing::debug!("Kernel/Initrd being used by other deployments. Won't delete");
-                }
-
                 break;
             }
 
@@ -103,72 +90,10 @@ fn delete_type1_entry(depl: &DeploymentEntry, boot_dir: &Dir, deleting_staged: b
             "Deleting staged entries directory: {}",
             TYPE1_ENT_PATH_STAGED
         );
+
         boot_dir
             .remove_dir_all(TYPE1_ENT_PATH_STAGED)
             .context("Removing staged entries dir")?;
-    }
-
-    Ok(())
-}
-
-#[fn_error_context::context("Deleting kernel and initrd")]
-fn delete_kernel_initrd(bls_config: &BLSConfigType, boot_dir: &Dir) -> Result<()> {
-    let BLSConfigType::NonEFI { linux, initrd, .. } = bls_config else {
-        anyhow::bail!("Found EFI config")
-    };
-
-    // "linux" and "initrd" are relative to the boot_dir in our config files
-    tracing::debug!("Deleting kernel: {:?}", linux);
-    boot_dir
-        .remove_file(linux)
-        .with_context(|| format!("Removing {linux:?}"))?;
-
-    for ird in initrd {
-        tracing::debug!("Deleting initrd: {:?}", ird);
-        boot_dir
-            .remove_file(ird)
-            .with_context(|| format!("Removing {ird:?}"))?;
-    }
-
-    // Remove the directory if it's empty
-    //
-    // This shouldn't ever error as we'll never have these in root
-    let dir = linux
-        .parent()
-        .ok_or_else(|| anyhow::anyhow!("Bad path for vmlinuz {linux}"))?;
-
-    let kernel_parent_dir = boot_dir.open_dir(&dir)?;
-
-    if kernel_parent_dir.entries().iter().len() == 0 {
-        // We don't have anything other than kernel and initrd in this directory for now
-        // So this directory should *always* be empty, for now at least
-        tracing::debug!("Deleting empty kernel directory: {:?}", dir);
-        kernel_parent_dir.remove_open_dir()?;
-    };
-
-    Ok(())
-}
-
-/// Deletes the UKI `uki_id` and any addons specific to it
-#[fn_error_context::context("Deleting UKI and UKI addons {uki_id}")]
-fn delete_uki(uki_id: &str, esp_mnt: &Dir) -> Result<()> {
-    // TODO: We don't delete global addons here
-    let ukis = esp_mnt.open_dir(SYSTEMD_UKI_DIR)?;
-
-    for entry in ukis.entries_utf8()? {
-        let entry = entry?;
-        let entry_name = entry.file_name()?;
-
-        // The actual UKI PE binary
-        if entry_name == format!("{}{}", uki_id, EFI_EXT) {
-            tracing::debug!("Deleting UKI: {}", entry_name);
-            entry.remove_file().context("Deleting UKI")?;
-        } else if entry_name == format!("{}{}", uki_id, EFI_ADDON_DIR_EXT) {
-            // Addons dir
-            tracing::debug!("Deleting UKI addons directory: {}", entry_name);
-            ukis.remove_dir_all(entry_name)
-                .context("Deleting UKI addons dir")?;
-        }
     }
 
     Ok(())
@@ -209,6 +134,9 @@ fn remove_grub_menucfg_entry(id: &str, boot_dir: &Dir, deleting_staged: bool) ->
     rename_exchange_user_cfg(&grub_dir)
 }
 
+/// Deletes the .conf files in case for systemd-boot and Type1 bootloader entries for Grub
+/// or removes the corresponding menuentry from Grub's user.cfg in case for grub UKI
+/// Does not delete the actual boot binaries
 #[fn_error_context::context("Deleting boot entries for deployment {}", deployment.deployment.verity)]
 fn delete_depl_boot_entries(
     deployment: &DeploymentEntry,
@@ -219,27 +147,15 @@ fn delete_depl_boot_entries(
 
     match deployment.deployment.bootloader {
         Bootloader::Grub => match deployment.deployment.boot_type {
-            BootType::Bls => delete_type1_entry(deployment, boot_dir, deleting_staged),
-
+            BootType::Bls => delete_type1_conf_file(deployment, boot_dir, deleting_staged),
             BootType::Uki => {
-                let esp = storage
-                    .esp
-                    .as_ref()
-                    .ok_or_else(|| anyhow::anyhow!("ESP not found"))?;
-
-                remove_grub_menucfg_entry(
-                    &deployment.deployment.verity,
-                    boot_dir,
-                    deleting_staged,
-                )?;
-
-                delete_uki(&deployment.deployment.verity, &esp.fd)
+                remove_grub_menucfg_entry(&deployment.deployment.verity, boot_dir, deleting_staged)
             }
         },
 
         Bootloader::Systemd => {
             // For Systemd UKI as well, we use .conf files
-            delete_type1_entry(deployment, boot_dir, deleting_staged)
+            delete_type1_conf_file(deployment, boot_dir, deleting_staged)
         }
 
         Bootloader::None => unreachable!("Checked at install time"),
@@ -249,31 +165,29 @@ fn delete_depl_boot_entries(
 #[fn_error_context::context("Deleting image for deployment {}", deployment_id)]
 pub(crate) fn delete_image(sysroot: &Dir, deployment_id: &str, dry_run: bool) -> Result<()> {
     let img_path = Path::new("composefs").join("images").join(deployment_id);
+    tracing::debug!("Deleting EROFS image: {:?}", img_path);
 
-    tracing::debug!("Will delete EROFS image: {:?}", img_path);
-
-    if !dry_run {
-        sysroot
-            .remove_file(&img_path)
-            .context("Deleting EROFS image")?;
+    if dry_run {
+        return Ok(());
     }
 
-    Ok(())
+    sysroot
+        .remove_file(&img_path)
+        .context("Deleting EROFS image")
 }
 
 #[fn_error_context::context("Deleting state directory for deployment {}", deployment_id)]
 pub(crate) fn delete_state_dir(sysroot: &Dir, deployment_id: &str, dry_run: bool) -> Result<()> {
     let state_dir = Path::new(STATE_DIR_RELATIVE).join(deployment_id);
+    tracing::debug!("Deleting state directory: {:?}", state_dir);
 
-    tracing::debug!("Will delete state directory: {:?}", state_dir);
-
-    if !dry_run {
-        sysroot
-            .remove_dir_all(&state_dir)
-            .with_context(|| format!("Removing dir {state_dir:?}"))?;
+    if dry_run {
+        return Ok(());
     }
 
-    Ok(())
+    sysroot
+        .remove_dir_all(&state_dir)
+        .with_context(|| format!("Removing dir {state_dir:?}"))
 }
 
 #[fn_error_context::context("Deleting staged deployment")]
@@ -293,9 +207,9 @@ pub(crate) fn delete_staged(
     }
 
     let file = Path::new(COMPOSEFS_TRANSIENT_STATE_DIR).join(COMPOSEFS_STAGED_DEPLOYMENT_FNAME);
-    tracing::debug!("Deleting staged deployment file: {file:?}");
 
-    if !dry_run {
+    if !dry_run && file.exists() {
+        tracing::debug!("Deleting staged deployment file: {file:?}");
         std::fs::remove_file(file).context("Removing staged file")?;
     }
 
@@ -307,6 +221,7 @@ pub(crate) async fn delete_composefs_deployment(
     deployment_id: &str,
     storage: &Storage,
     booted_cfs: &BootedComposefs,
+    dry_run: bool,
 ) -> Result<()> {
     const COMPOSEFS_DELETE_JOURNAL_ID: &str = "2a1f0e9d8c7b6a5f4e3d2c1b0a9f8e7d6";
 
@@ -357,11 +272,11 @@ pub(crate) async fn delete_composefs_deployment(
         ""
     };
 
-    tracing::debug!("Deleting {kind}deployment '{deployment_id}'");
+    tracing::info!("Deleting {kind}deployment '{deployment_id}'");
 
     delete_depl_boot_entries(&depl_to_del, &storage, deleting_staged)?;
 
-    composefs_gc(storage, booted_cfs, false).await?;
+    composefs_gc(storage, booted_cfs, dry_run).await?;
 
     Ok(())
 }
